requirements.txt
asgiref==3.7.2
Bottleneck==1.3.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.11
django-import-export==3.3.7
django-mathfilters==1.0.0
django-session-timeout==0.1.0
et-xmlfile==1.1.0
flake8==7.0.0
flake8-html==0.4.3
Jinja2==3.1.3
llvmlite==0.42.0
MarkupPy==1.14
MarkupSafe==2.1.5
mccabe==0.7.0
numba==0.59.1
numexpr==2.10.0
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.1
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
python-calamine==0.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyxlsb==1.0.10
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.10.0
tzdata==2024.1
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
